Move sbatch model to models dir
